requirements.txt
boto3>=1.34.0
botocore>=1.34.0
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-asyncio>=0.21.0
moto>=4.2.0
tabulate>=0.9.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
flask>=2.3.0
flask-cors>=4.0.0
gunicorn
google-generativeai
openai
